Feature: Transactions and BatchWrite operation classes (#291)
* feat: add Dyngoose.Transaction and Dyngoose.BatchWrite operation classes

Extends the work by @priyanshu966106 into a utility class which I hope
will make the functionality a bit more robust. The original
implementation works as expected, but I spent some time thinking about
the implementation of the batch write and this new transact write. Part
of the usefulness of these within DynamoDB is that you can perform a
batch write of several documents from different tables, which can be
useful if you are creating a several types of records at once.

This new transact write method makes usefulness of being able to write
across several tables at once far more useful and it would be nice to
expose that. An operation could touch several tables; and you want the
entire operation to work successfully or fail. We can expose that with
`transactWrite` where the process is atomic, if any part fails the
entire operation will have no affect. This is a very useful feature, but
it is really more useful if it was exposed in a way that worked across
several tables.

So I worked on an updated proposal for this, which works using a class
that allows you to build I'm thinking it might be useful to expose new
classes globally, like:

```typescript
const transaction = new Dyngoose.Transaction()

// Dyngoose will use Put or Update as necessary
transaction.save(item)

// you could use .put or .update explicitly to be more discrete
transaction.put(item)

// adds a delete request
transaction.delete(item)

// adds a condition to an item not being updated
transaction.conditionCheck(item, conditions)
await transaction.commit()
```

Similarly, for batch write operations, you could use:

```typescript
const batch = new Dyngoose.BatchWrite()
batch.put(item)
batch.delete(item)
await batch.commit()
```

It might be useful to do something similar with the BatchGet operation,
which also supports reading from several tables at once; but I think the
use case is a bit more limited and the implementation would be a bit
difficult to deal with since the response would contain a mix and the
order isn't guaranteed. As it stands, the current `batchGet` operation
does not handle the `UnprocessedKeys` correctly. So that will come later.

* docs: add docs for transactions