feat: remove moment in favor of using native Date functions (#292)
Since momentjs published the library will no longer be in active
development, it made me look at whether Dyngoose really needed to use
momentjs for anything. The truth is that it did not.

The @Dyngoose.Attribute.Moment() was not being exposed like other
attribute types and the only other use of moment was in the Date
attribute for handling some really basic formatting and parsing, all of
which is possible using native Date functions and is quite reliable
within Node.js.