feat: set limit of 1 when performing a gsi.get query
this changes the query to become more performant, and added additional
checks to throw errors if the .get method is being used for unintended
purposes. if using .get for anything other than a .primaryKey.get like
operation, you should be using .query and handle the situation when
multiple records are returned.