feat: atomic batch get operations
Added support for TransactGetItem operations, however, I've added this
to the existing BatchGet class because the usage is so similar. To
convert your operation to an atomic transaction, simply call `.atomic()`
and internally Dyngoose will use TransactGetItem instead of a BatchGet.

Also converted the PrimaryKey.batchGet to consume the BatchGet class,
this removes some redundant code but it also deprecated the batchGet
method as it does not expose all the same functionality available via
the BatchGet class.