fix: do not apply defaults when loading from DynamoDB
This is to resolve a bug, where attributes using defaults, which
includes dates using nowOnCreate and nowOnUpdate. This is a problem if
you use a projection to load only part of a record, it would reset the
attributes not loaded back to their defaults which can lead to updating
attributes you do not intend to update or to reading incorrect values.

While it might be possible to look at the projection and understand
which attributes did not get requested, that would require a lot of
changes to the way the instances of tables are created.

This solution created a new `applyDefaults` method on the table which
is automatically called when using the `Table.new()` typed method. Since
that is really the ideal way to create new records, this should be fine.

This might "break" some uses of Dyngoose where they're using
`new Table()` and expecting the defaults to get applied, however, the
bugs this resolves are important.