feat: throw helpful errors with useful stacks
Errors still inherit and confirm to the AWSError object, but include an
abbreviated stack trace that terminates at the Dyngoose methods. Errors
also now includes the table name, table class object, and the query
or scan input called for diagnostics. The query input is not included
in the error message, as that can easily lead to security concerns, but
including it on the error object might prove helpful when debugging
tricky query issues.