refactor: rename `delete` attribute and property methods to `remove`
I think this was a poor name choice on my part, using `del` to delete
a property, `deleteAttribute`, and `deleteAttributes`, but then to
delete the record is the `delete` method.

This renames everything to `remove`, so `remove` removes a property,
and `deleteAttribute` is now `removeAttribute`.

Keeps backwards-compatibility to make upgrading to Dyngoose 3.0.0
easier