feat: utility to update sets more intelligently (#570)
Dyngoose tries keep track of "changed" properties, so when saving, it
can intelligently perform an UpdateItem operation. Sets pose some
complications. Sets in Dyngoose do not retain the order of the items
within the set and they cannot contain duplicates. So this PR adds a
utility function to update sets intelligently.

I've attempted to add strict typing, however, TypeScript compiler fails
to understand the complexity so I had to cop out a bit.