fix(date): passing invalid date values to fromJSON would error
Using the Table.fromJSON method would fail when passing empty values
such as null, undefined, or an empty string because the fromJSON method
was implementing expecting a non-empty value. Changed the logic within
the Table class to only transform the value using fromJSON after we
verify the value being provided is not empty.