fix: map attributes using independent property names
See #600. Most of this fix was pulled from @cbumstead's work, however,
I kept the attributes map keyed by the attribute name and fixed the
map object build logic in toDynamo to use the attribute name while
ensuring that we do not leave lingering values on the map in fromDynamo.

There was a previous side-effect, that map properties loaded from dynamo
not defined as attributes would be returned in their Dynamo
AttributeValue format without being manipulated. This side effect has
been removed, although the unintentional value was that it allowed for
easy reading of those arbitrary attribute values. Instead, the next
commit will add explicit support for arbitrary values on maps.