fix: migrateTable comparison of indexes
Looks like aws dynamodb client v3 always provides a default property
on the NonKeyAttributes of undefined, which is breaking the comparison
logic using _.isEqual to determine if the project has changed. Updated
to ensure we always have a default value for both objects set to ensure
comparison does not break when nothing has changed.

Fixes #645