feat: abort controller support
Refactored away from a separate requestOptions object to utilize the
existing input options object for all the read request methods. I think
this will make for an simpler implementation in use.

Additionally, added a mention to the README about this feature.

Lastly, actually trigger a release to get this out.