fix: using a conditional check on a nowOnUpdate Date attribute
We need to remove the legacy logic for trying to ensure we always update
a Date with "nowOnUpdate" enabled. We now have a method on the table
called "updateOnSaveAttributes" that ensures properties that need to
self-check themselves do so, so this check is not necessary. By removing
this, it allows you to use a conditional check comparison (such as in a
transaction) against the date's previous value.